pyproject.toml
[tool.poetry]
name = "electronic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.6"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
drf-yasg = "^1.21.10"
djangorestframework-simplejwt = "^5.5.1"
django-filter = "^25.1"


[tool.poetry.group.lint.dependencies]
black = "^25.9.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
